# src/main/java/com/example/library/dto/response/UserRegisterResponseDto.java
package com.example.library.dto.response;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.Setter;

@EqualsAndHashCode(callSuper = true)
@Data
@Setter
@JsonInclude(JsonInclude.Include.NON_NULL)
public class UserRegisterResponseDto extends UserResponseDto {

    private String message;
}